DRAFT: Remove indirect dependencies on std.
The primary change here is to disable default features on `rand` and
`num-traits`, so that those libraries do not depend on `std` themselves.
Then, it becomes necessary to remove dependencies on parts of
`num-traits` that don't exist without the std feature, mainly `Float`.

This commit is not complete; I'm committing a WIP for someone else to
consider taking over.